the thread sanitizer reports a memory use after free problem when runnint the two_renames_race test. it may take several (<1000) test runs to reproduce this problem.  compile with clang with the thread sanitizer enabled.
the problem occurs because of a race on the m_todo data structure.  the old code manipulates the m_todo
as a stack (in a vector).  there is a problem in the copier that peeks at the back, does some work, and
then pops the back.  unfortunately, another thread could have pushed another item onto the m_todo stack
in between.

perhaps using a deque and adding to the back and removing from the front is a better data structure.
at least the race is fixed.